UPDATE : Move init fetchUrl into App.js
