+++ front/Mobile/src/Login/Login.js
@@ -4,41 +4,18 @@ import { useFocusEffect, useNavigation } from '@react-navigation/native';
 import { useState, useCallback } from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import NetworkLocation from '../NetworkLocation/NetworkLocation';
+import { getFetchUrl } from '../getFetchUrl';
 
 export default function Login() {
   const navigation = useNavigation();
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [fetchUrl, setFetchUrl] = useState('');
 
   useFocusEffect(
     useCallback(() => {
-      const storeFetchUrl = async () => {
-        try {
-          const url = await AsyncStorage.getItem('fetchUrl');
-          if (url === null) {
-            await AsyncStorage.setItem('fetchUrl', 'http://inox-qcb.fr:8000');
-            setFetchUrl('http://inox-qcb.fr:8000');
-          } else {
-            setFetchUrl(url);
-          }
-        } catch (error) {
-          console.error('Error:', error);
-        }
-      };
-
-      storeFetchUrl();
-    }, [])
-  );
-
-  useFocusEffect(
-    useCallback(() => {
-      if (fetchUrl === '') {
-        return;
-      }
-
       const checkLoginStatus = async () => {
         try {
+          const fetchUrl = await getFetchUrl();
           const token = await AsyncStorage.getItem('accessToken');
           if (token === null) {
             return;
@@ -57,11 +34,12 @@ export default function Login() {
         }
       };
       checkLoginStatus();
-    }, [navigation, fetchUrl])
+    }, [navigation])
   );
 
   async function onLogin() {
     try {
+      const fetchUrl = await getFetchUrl();
       const response = await fetch(fetchUrl + "/login", {
         method: "POST",
         headers: {
